refactor: split types.go into 7 domain-focused files (Batch G)
50 type definitions split by domain, all in package core:
- config_types.go (19 types): HooksConfig, CoachingConfig, etc.
- dispatch_types.go (3): HookPayload, DispatchResult, HandlerResult
- guard_types.go (3): GuardRuleConfig, GuardResult, ParsedCondition
- feed_types.go (10): CacheEntry, FeedDefinition, feed configs
- analytics_types.go (6): AnalyticsEvent, SessionSummary, etc.
- handler_types.go (3): ResolvedHandler, SegmentConfig, etc.
- types.go (6): Event constants, ValidEventType, test types

No exported API changes. All consumers see identical symbols.
